backend/requirements.txt
# Core Flask
Flask==3.0.3

# CORS support for React frontend
Flask-CORS==4.0.1

# Environment variable management
python-dotenv==1.0.1

# Production WSGI server
gunicorn==22.0.0

# Testing
pytest==8.2.2
pytest-flask==1.3.0